#3101: Move metal_BERT_large_15 to models folder
