+++ models/experimental/metal_BERT_large_15/tests/test_bert_encoder.py
@@ -1,0 +1,152 @@
+# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import pytest
+import torch
+
+from transformers import BertForQuestionAnswering
+from loguru import logger
+
+
+import tt_lib
+
+from tt_lib.utils import pad_activation
+from models.utility_functions import comp_pcc, comp_allclose, profiler
+from models.experimental.metal_BERT_large_15.tt.model_config import get_model_config
+from models.experimental.metal_BERT_large_15.tt.bert_encoder import TtBertEncoder
+
+class PytorchBertEncoder(torch.nn.Module):
+    def __init__(self, hugging_face_reference_model):
+        super().__init__()
+        self.bert_encoder = hugging_face_reference_model.bert.encoder.layer[0]
+
+    def forward(self, x, attention_mask=None):
+        return self.bert_encoder(x, attention_mask)[0]
+
+def run_bert_encoder_inference(
+    device, model_version, batch, seq_len, pcc, model_config, model_location_generator
+):
+    model_name = str(model_location_generator(model_version, model_subdir="Bert"))
+
+    hugging_face_reference_model = BertForQuestionAnswering.from_pretrained(
+        model_name, torchscript=False
+    )
+    config = hugging_face_reference_model.config
+
+    tt_bert_encoder_model = TtBertEncoder(
+        hugging_face_reference_model.config,
+        0,
+        hugging_face_reference_model.state_dict(),
+        device,
+        model_config,
+    )
+    pytorch_bert_model = PytorchBertEncoder(hugging_face_reference_model)
+
+    # Prepare input
+    torch.manual_seed(0)
+    bert_encoder_input = (
+        torch.rand(batch, 1, seq_len, hugging_face_reference_model.config.hidden_size)
+        * 2
+    ) - 1
+    bert_attention_mask = torch.zeros(batch, 1, 1, seq_len)
+    extended_bert_attention_mask = torch.zeros(batch, 1, 32, seq_len)
+
+    pytorch_out = pytorch_bert_model(
+        bert_encoder_input.squeeze(1), bert_attention_mask
+    ).unsqueeze(1)
+
+    pad_bert_encoder_input = pad_activation(bert_encoder_input)
+    tt_bert_encoder_input = (
+        tt_lib.tensor.Tensor(
+            pad_bert_encoder_input.reshape(-1).tolist(),
+            bert_encoder_input.shape,
+            model_config["OP1_FUSED_QKV_MM_INPUT_DTYPE"],
+            tt_lib.tensor.Layout.ROW_MAJOR,
+        )
+        .to(tt_lib.tensor.Layout.TILE)
+        .to(device, model_config["OP1_FUSED_QKV_MM_INPUT_MEMCFG"])
+    )
+    tt_bert_attention_mask = (
+        tt_lib.tensor.Tensor(
+            extended_bert_attention_mask.reshape(-1).tolist(),
+            extended_bert_attention_mask.shape,
+            model_config["OP8_SOFTMAX_ATTENTION_MASK_DTYPE"],
+            tt_lib.tensor.Layout.ROW_MAJOR,
+        )
+        .to(tt_lib.tensor.Layout.TILE)
+        .to(device, model_config["OP8_SOFTMAX_ATTENTION_MASK_MEMCFG"])
+    )
+
+    tt_out = tt_bert_encoder_model(tt_bert_encoder_input, tt_bert_attention_mask).cpu()
+
+    tt_out = tt_out.to(tt_lib.tensor.Layout.ROW_MAJOR).to_torch().reshape(tt_out.shape())
+
+    passing, output = comp_pcc(pytorch_out, tt_out, pcc)
+    logger.info(f"Output {output}")
+
+    _, output = comp_allclose(
+        pytorch_out, tt_out, 0.5, 0.5
+    )  # Only interested in reporting atol/rtol, using PCC for pass/fail
+    logger.info(f"Output {output}")
+
+    if not passing:
+        logger.error(f"Output PCC < {pcc}")
+
+    if model_config["DEFAULT_DTYPE"] == tt_lib.tensor.DataType.BFLOAT8_B and not passing:
+        pytest.xfail("PCC is garbage for BFLOAT8_B. Numbers are for perf only!")
+
+    assert passing
+
+
+@pytest.mark.parametrize(
+    "batch, model_config_str",
+    (
+        (9, "BFLOAT8_B-DRAM"),
+        (9, "BFLOAT16-DRAM"),
+        (9, "BFLOAT8_B-L1"),
+        (9, "BFLOAT16-L1"),
+        (9, "MIXED_PRECISION_BATCH9"),
+        (8, "MIXED_PRECISION_BATCH8"),
+    ),
+    ids=[
+        "batch_9-BFLOAT8_B-DRAM",
+        "batch_9-BFLOAT16-DRAM",
+        "batch_9-BFLOAT8_B-L1",
+        "batch_9-BFLOAT16-L1",
+        "batch_9-MIXED_PRECISION_BATCH9",
+        "batch_8-MIXED_PRECISION_BATCH8",
+    ],
+)
+@pytest.mark.parametrize(
+    "model_version, seq_len, pcc",
+    (("phiyodr/bert-large-finetuned-squad2", 384, 0.99),),
+    ids=["BERT_LARGE"],
+)
+def test_bert_encoder_inference(
+    model_version,
+    batch,
+    seq_len,
+    pcc,
+    model_config_str,
+    model_location_generator,
+    request,
+    device
+):
+    model_config = get_model_config(model_config_str)
+
+    tt_lib.profiler.set_profiler_location(
+        f"tt_metal/tools/profiler/logs/BERT_large_1_encoder_{request.node.callspec.id}"
+    )
+
+    tt_lib.profiler.start_profiling("entire_run")
+    run_bert_encoder_inference(
+        device,
+        model_version,
+        batch,
+        seq_len,
+        pcc,
+        model_config,
+        model_location_generator,
+    )
+    tt_lib.profiler.stop_profiling("entire_run")
